Run the flow-map target in fp32, and pin two bf16-fragile guards
one_step_euler_loss computes logmap(x0, x_end) / step: it differences two O(1)
states and divides by a step as small as 2**-(sections-1). That turns bf16's
~2**-8 absolute rounding into 2**-8/step relative error -- measured 47% at
d=2**-7 Euclidean, and 100% on the sphere factors by d=2**-5, since arccos is
ill-conditioned for near-parallel vectors. x_end comes from two no-grad
forwards, so it is corrupted supervision rather than gradient noise and no
amount of training removes it.

Measured, 600 steps resumed from had_lognorm_v2, everything else identical:

  precision              out_dir  out_pos   energy  one_step_euler   time
  fp32 everywhere        0.00367  0.00271  0.01047      0.00007      344s
  bf16, unguarded        0.00507  0.00415  0.01105      0.02716      288s
  bf16 + this guard      0.00372  0.00281  0.01062      0.00007      305s

The guard recovers fp32's flow map exactly (1.018x vs 389x unguarded) at 3.7%
on the other channels, and stays 11% faster than fp32 everywhere.

Only this function needs it. The solvers accumulate small increments into an
O(1) state, which keeps ~2**-8 relative error instead of amplifying it; bf16
generation measures within 0.98-1.07x of fp32 at every step size.

Also pins two elementwise guards that are correct today only because their
inputs happen to arrive as fp32 -- autocast casts matmuls, not elementwise math:
base_nn's logit(clamp(1e-4, 1 - 1e-4)), where 1 - 1e-4 rounds to exactly 1.0 in
bf16 and logit(1.0) is +inf; and _ot_couple's clamp(-1 + 1e-6, 1 - 1e-6).acos(),
where the same rounding silently removes the guard at the worst-conditioned end.

This is why every model that trained a flow map ran precision 32 and every bf16
model had one_step_euler_fac 0 -- had_distill_070926 and had_lognorm v1/v2 are
the only four runs that combined them, and the only four that failed.

+++ legofmt/base_dist/base_nn.py
@@ -119,7 +119,11 @@ class BaseDist:
 
         ed_t = ds_t.f.edep
         w    = fwd.float() * (ed_t > 0)
-        ly   = torch.logit((ed_t / self.gen_base.e_dep_max).clamp(1e-4, 1 - 1e-4))
+        # .float() is load-bearing: in bf16 `1 - 1e-4` rounds to exactly 1.0
+        # (resolution near 1 is 2**-8), so the clamp stops bounding and
+        # logit(1.0) is +inf. Safe today only because ed_t comes from fp32 data
+        # -- autocast casts matmuls, not elementwise math -- so pin it.
+        ly   = torch.logit((ed_t.float() / self.gen_base.e_dep_max).clamp(1e-4, 1 - 1e-4))
         mu1, sig1 = mu.squeeze(-1), sig.squeeze(-1)
         l_edep = ((sig1.log() + (ly - mu1) ** 2 / (2 * sig1 ** 2))
                   * w).sum() / w.sum().clamp(min=1)
@@ -160,9 +164,13 @@ class BaseDist:
         man = self.rc.manifold
         tgt = ds_t.f.out_cc.unsqueeze(-2).split(man.ambient_dims, dim=-1)
         ref = out.unsqueeze(-3).split(man.ambient_dims, dim=-1)
+        # .float() is load-bearing on the sphere branch: in bf16 `1 - 1e-6`
+        # rounds to exactly 1.0, so the clamp silently stops guarding acos at
+        # the near-parallel end where it is worst conditioned. `ref` derives
+        # from base_head's Linear layers, which autocast DOES cast.
         cost = sum(
-            ((a * b).sum(-1).clamp(-1 + 1e-6, 1 - 1e-6).acos()
-             if isinstance(mf, Sphere) else (a - b).norm(dim=-1)) ** 2
+            (((a * b).sum(-1).float().clamp(-1 + 1e-6, 1 - 1e-6).acos()
+              if isinstance(mf, Sphere) else (a - b).norm(dim=-1).float()) ** 2)
             for mf, a, b in zip(man.manifolds, tgt, ref)
         ) + blocked * 1e6
         assign = slap(cost, cost.device).long()
